corrected basecommand for running grpc
