+++ docker-compose.dev.yml
@@ -14,6 +14,8 @@ services:
     environment:
       PYTHONWARNINGS: 'ignore'
     restart: always
+    command: >
+      bash -c "python3 manage.py run_grpc_server '[::]:50051'"
     working_dir: /iam
     ports:
       - '50051:50051'
