+++ sgb/run_grpc_server.sh
@@ -1,0 +1,1 @@
+#!/usr/bin/env bash
